Rename lotIsClosed helper file
