+++ src/v2/Apps/Artwork/Utils/lotIsClosed.ts
